điều chịnh lại bán hàng tại quầy và bán lẻ

+++ src/main/java/com/example/Trendora/DuAn/DTO/HoaDonChiTietDTO.java
@@ -16,9 +16,9 @@ public class HoaDonChiTietDTO {
     private Integer idHoaDon;
     private String maHdct;
     private String tenSanPham;
-    private Integer donGia;
+    private BigDecimal donGia;
     private Integer soLuong;
-    private Integer thanhTien;
+    private BigDecimal thanhTien;
     private Integer trangThai;
 }
