+++ src/main/java/com/example/Trendora/DuAn/model/HoaDonChiTiet.java
@@ -32,13 +32,13 @@ public class HoaDonChiTiet {
     private String maHdct;
 
     @Column(name = "don_gia")
-    private Integer donGia;
+    private BigDecimal  donGia;
 
     @Column(name = "so_luong")
     private Integer soLuong;
 
     @Column(name = "thanh_tien")
-    private Integer thanhTien;
+    private BigDecimal thanhTien;
 
     @Column(name = "trang_thai")
     private Integer trangThai;
